Skeletons are automatically added, thus the "Add Skeletons" button has been  removed
Tweaks - Now, deleting an exercise will immediately remove it from the exercise list, rather than needing the page to reload

Notes:  Relying on the model->save() function on the server side to initialize variables (such as ID) is unreliable (i.e., doesn't work).  Is the mysql query asynchronous, or what is going on there?