This should be better.  Here's what was happening:
You can see in the code that curPath is updated every time a tab is selected that is not the fileBrowser.  If you were moving from one non-browser tab to another, the selectedItem stayed the same.  Each time you returned to the fileBrowser, the selected item is reset to nothing (null), even if it loaded to the path of the previously selected item (which it should, and would, do).

Before Dusenberry made it more efficient, when selecting another tab the reloading into root was happening behind the scenes, so when you came back to the fileBrowser you would be at the root level, but you wouldn't see that flicker of the expanded directory before reloading to root.

After r443, you would see the flicker because the reloading was happening only behind the scenes - the flicker making it more obvious that we were reloading into the root directory after having it expanded.

r444 fixed the flicker, by breaking reloading.  r445 fixes the flicker like it should be fixed (I believe).