Descriptions should be working now (I think), with a few minor issues:
When uploading descriptions, an empty red bar appears at the top as though there was an error. I can't figure out how to fix that, and it's making me feel incompetent.

We need to create a link on the live version from ..../descriptions/desc to /tmp for this to work.