Client side changes to coincide with previous commit.
Philip: I attempted several times to do what you suggested, but I was having issues getting it to work. When I called explode() on segments, innerfunctions, booleans, etc. it would just return an array with the elements separated by ",". This didn't really work out, since that could easily exist within code.

I decided to just go with what I had before, as it works and is a lot simpler. Now it just passes the string through to the client, and the client splits on .:|:. -- you can check parseText() in WEStatus. I left your implementation in there, just in case you decide you want to do something with it. I have a feeling that whatever you were doing in getData() in MagnetProblem.php was returning an array of arrays, so you could parse that with what you were doing. I fiddled with stuff for a while to get it to work, but I failed.

I was going to remove getData() from that class, but I felt really bad because it was so beautiful. I think you might want it removed, but I feel bad killing your creation.

I ran my little script to modify the database, so it should be working now. I made a backup immediately before running it, though, just in case it broke everything.