RadixSort looks like it is working
Had to do all sorts of shuffling and passing that isn't really the best, and may hamper additional functionality of SearchProblem, as it now will only take Evaluation_RadixSortWithHelp as its eval.  That's all we were wanting to give it, so for now it seems to be fine.

However, any time a magnet is dropped we get an umbrella exception, so it looks like the grid drop controller may not be overly happy with us.  But it is working.