Automatically nesting comments.
Copied from the email I sent out:

We now have automatically nesting comments for the innerFunctions and premadeSegments(statements as referred to in the database).
This will allow you to have comments start out inside those magnets.  I believe this could be used to provide some extra guidance as to the order of the steps needed to make an algorithm, I'm not terribly sure what the expectations for it's use are.

Anyway, how to use this.  So right now you have your given magnets setup like:

magnet1.:|:.magnet2.:|:.magnet3

For comments, after each magnet but before the delimiter you'll have a second delimiter, ".:2:.", followed by your comment like this:

magnet1.:2:.comment1 for magnet1.:2:.comment2 for magnet1.:|:.magnet2.:|:.magnet3.:2:.comment1 for magnet3.:2:.comment2 for magnet3

Now right now I have it setup so that it's just making a single line comment, "// comment", so I made it so you just have to put plaintext for the comments, no html tags or anything fancy.