This commit was a collaborative effort between me and Daniel. <3
The Comprehensive Review button under exercises in the editor should now include magnet problem submissions. I think it should work, but it was kind of broken because Admin was showing up for MagnetSubmission::getSubmissionsById(), and the code was not expecting to see Admin.

Philip - is there a way you can test this so that it isn't with section 19? I think that was messing up our tests.