Rework of problem creation UI.  Still looks pretty bad, but at least is kind of organized how I'd like to see it when all polished.
For some reason, I can't get the title, description, and comment textboxes/area to go to 100% width.  When I change it in the browser it looks like what I want, but I can't get the ui.xml file to initialize it with that.

Regardless, this is just a precursor to actually implementing all of the problem creation server side stuff.  Things on the client should work like they did before, although some functionality which is implied by the UI won't currently be there as I added some things to the UI, and didn't just move things around.