When attempting to save a problem with a name that is already used, IF you were the creator of that problem, you are asked if you want to overwrite it.
If not, you get to change the title.

If so, the problem gets overwritten (not deleted - so if it was assigned before, it is still assigned now, etc).  HOWEVER, if you do not supply any test files, then the old test files remain (in case you were just fixing a typo).  BUT, if you supply any additional files, ALL the other files will get deleted and replaced with whatever you add (in case you are creating a completely new problem with the same name).