A whole series of changes as we start moving LogicalMicrolabs into the database.  There are a few things I am not sure of, stylistically speaking.
One is, if you go into DSTConstants you'll see I added some methods that maps the enums in the database to the evaluation/edgerules classes, and that they are if chains.  I could implement this is as a HashMap, but it's not like the code will get any more readable as we'll just have a long chain of adding things to the HashMap.  I'm also not sure it would actually run faster, as I'm not sure if the HashMap would be written every time, or how to stop that from happening.

Another is, in the DB I'm using enums for a lot of categories.  I like enums as when we have to work in the database we won't have to keep referencing some map to figure out what each integer means.  But, I'm not sure about the general etiquette for using enums.  Any of you DB folk - what is the general consensus on that?

In short, this is how this works:  Unfortunately, you can't really have half the microlabs get accessed by the DB and the other half by ProblemServiceImpl because of the asnychronous nature of requests - you end up getting strange hybrid microlabs as you get what looks like two threads running at the same time in the client.

So, the "switch" point is in DataStructureTool.getProblem() ->  either calls ProblemServiceImpl.getProblem and then initialize(), or calls Proxy.getLogicalMicrolab() which then calls initialize itself.

Communication between the client and the server happens through getLogicalMicrolab on the server which returns an array (similar to getMagnetProblem) that we use on the client to build a LogicalMicrolab object.  One parameter is the "genre," which will probably change, but the genre corresponds to the groups on the excel sheets. 

The LogicalMicrolab was a "getProblem()