Finally trying to commit. This may break Everything. - Client side commit for saving created magnets (It looks like Philip has already committed my server side stuff)
Most of the changes are simply passing down an extra parameter I had to bring in or checking that parameter to see if a magnet is one of the created magnets or not.

Also have saveMagnetState taking a boolean for whether the save is coming from the button being pressed or when they get the problem right.  This is to be able to only display the "State Saved" notification when the click the button.