Added a lot of functionality to the Logical Admin page, but the flickering is going to drive me insane.
The two biggest flickering issues are 1)  When it first loads (looks real bad, random scroll bars popping up, etc) and 2) on occasion when changing pretty much anything stuff slightly scoots to the left and then scoots back to the right when things are loaded.

I'd love for someone to have a suggestion to fix this.

Also, be aware of a ProxyFacilitator interface, which is going to clean up Proxy for us.  Eventually.