Renamed creation station and continued refactoring.
Also, two bugs:

Hitting reset on a magnet maker problem will NOT reset the limits.

Limits do not increase recursively when deleting magnets. For example:

Create 5 if statements. Nest the last 4 inside of the first one. Delete the top level if. Notice that you now have only 1 if statement to use instead of 5.