Continued work on LMBuilding.
Main thing was re-creating the logical microlab canvas for instructors to use to build their exercises.  I decided against trying to decipher what was going on in logical microlabs currently and basically do it from the ground up.

So, how it happens now is basically this:  All parent-child relationships between nodes occur when edges are built.  There are no edge collections or node collections, but rather quite literally a binary search tree datastructure behind the scenes implemented as a tree of BasicNodes, which extend Label - tightly coupling the visual representation and the actual data.  Coupling which I am a huge fan of.

All node relationships occur when the edge is created, and are dismantled when an edge is removed.  If a node is moved, the previous edge relationship is destroyed and replaced by the new ones - all shown by the new edges that are built.

Thus, we'll be able to do a literal implementation of the traversal algorithms to build the necessary traversals, at least for all the binary search tree problems. 

The canvas is the "BasicBuilder" class.  May be renamed.  Heap and Graph exercises will probably extend this BasicBuilder, although honestly most of the current functionality should be universal.

To kind of give an overview:  LMBuilder.java is an aggregate of all the information necessary for creating a Logical Microlab in the database.  LMDisplay is the view that informs LMBuilder.  BasicBuilder (it is actually poorly named now that I think about it) is a part of LMDisplay that is used to construct the tree (and in the future heap or graph).

LMBuilder gets instantiated through LMBuilderFactory, as there are going to be a series of different LMBuilders with different default values depending on what type of Logical Microlab is being created.  Each group button decides what LMBuilder to create and what LMDisplay to use.  Each LMDisplay gets a reference to LMBuilder.

HOWEVER - looks like pulling logical microlabs down from the repo isn't actually working for all of them!  We can get them loaded and showing up, but not always correctly.  For example, heap building exercises don't give you the opportunity to add edges!  

BTW, to add edges in LogicalProblemCreation you just double click.  Yay.

Alright, that's enough of a comment.