DB - Arrow stuff mostly
This isn't much as I spent about 4 more hours working on Capstone than I expected.  But anyway, I just changed up the way the arrows draw a bit and tried to make it so that the whole page does not have to be redrawn when a DropPoint changes size.  Currently, if I try to remove and then redraw the arrows, everything breaks, but I will commit with that commented out and try some more tomorrow.