DB - Added the framework for undo functionality.
Added an ActionState abstract class that is used to push onto a stack.  The stack will hold state information for all previous commands that a user has performed.  The undoStack is private within FlowUI and can only be accessed via methods in that class.  This is so that the stack is in a single, accessible place and does not have to be passed around.  Though that can change if someone has a different preference.  In order to execute an undo, flow.popUndoStack().undo() should be called.

Instead of multiple constructors, each concrete ActionState should take whatever information is needed in order to restore its state and implement undo().  I know this doesn't really fall under a true Command Pattern, but this implementation should give extensibility for any ActionType that might come up in the future along with keeping individual classes clean/neat.