DB - adding actions for the basic operations.
Added executions/undo's for: Set, Add, Subtract, Multiply, Divide, and Mod.
I have an idea for what I'll have to do for Conditionals and Answers but

DropPoint and the DropControllers are looking pretty ugly right now but once I get a better idea of how everything is going to work out I'll do a major refactoring.  This is still more of the prototyping stage.  I'll probably end up with enums everywhere.