DB - Reset for Answer boxes now works. Variables in the SegmentsPanel now have the current value if any show up.
Also put all the actions into their own package because it was getting cluttered in there.

So those are two things off the list. However, I realized that I had left out the need for a TrashBin so I added that to the list.

TODO/Bugs:
- Add TrashBin
- Start setting up more thorough error checking/providing feedback.
- Split up SegmentPanel into distinct areas for the different types.
        - Add a DropHandler to the section that holds Answer_Choices as that is
the only one that should be able to be dragged back over since it doesn't drop a
copy of itself.
- Start getting it ready to pull down problems from the database
        - Figure out how we're going to store/parse the DropPoint types &
locations
        - Figure out how we're going to store what Segments they gets (includes
the variables, conditionals, answer_choices and other segments that require some
given content)
- Make it all look pretty