More refactoring wihle continuing to work towards MST problems.
Re-did how edges get redrawn to make a lot more sense.  

Due to the limitations of UiBinder, I rely on NodeHandlers and EdgeHandlers to change the functionality of a BasicCanvas.  Due to this structure, it is difficult to get edges to "remember" their weight when they get redrawn.  

As it's not a huge hassle to re-enter the weight and this automatically takes care of the "what if the prof wants to change the weight later" problem, I'm currently going to leave it as is.

Now, I mainly just have to learn how Kruskal and Prim actually run in code and implement that, and those two should be up.