Implementing a somewhat responsive ConstructUi area as far as layout goes.
The magnet groups will display in two columns if they have the room, otherwise they'll go down to one column.

Right now the resize is only really triggered when the slider is moved or when a magnet is moved.  So if the window gets resized then it won't be triggered until one of those two things happen.

To get it to actually tap into the window resize you would have to chain RequiresResize and ProvidesResize all the way up to the RootPanel.