Fixing the fact that the panels inside the magnets weren't showing up when trying to use the "Test Problem" button.
The reason was that the brackets we getting encoded at a different point in time.  Everything in the statements box was getting encoded instead of just the text that's going to show up in the label.

This should work. I tested it with a problem that had pseudocode and just labels and the problem appeared right and compiled and ran correectly.