Alright, this is more of a "repair" on my last commit.
basically what was (and still is) causing issue was me attempting to encode everything in the statements text area at once after the user clicked test Problem or commit. I was doing this thinking that they may have types something directly into the text area. I was having many issues with this which I somehow overlooked before that last commit.

As of now '"', '>', and '<' are only encoded when the user creates statements via the magnet builder feature on the left hadn side of the screen.