BIG BIG BIG COMMIT, PULL DOWN AT YOUR OWN RISK
This is the Current Version of what I have been working on to update the WAGS site. LOTS OF STUFF IS CHANGED. Neither of you are working on problems that are all that relevant to these changes currently so you can probably ignore most of these files. 

The beginnings of a new directory structure are being put in place, styles are being modified and gwt-bootstrap is being included. Send me any questions you have. I will work on documentation and such at a later date.